+++ ChangeLog.md
@@ -12,6 +12,7 @@
 * stack can act as a script interpreter (see [Script interpreter] (https://github.com/commercialhaskell/stack/wiki/Script-interpreter) and [Reddit discussion](http://www.reddit.com/r/haskell/comments/3bd66h/stack_runghc_turtle_as_haskell_script_solution/))
 * Add the __`--file-watch`__ flag to auto-rebuild on file changes [#113](https://github.com/commercialhaskell/stack/issues/113)
 * Rename `stack docker exec` to `stack exec --plain`
+* Add the `--skip-msys` flag [#377](https://github.com/commercialhaskell/stack/issues/377)
 
 ## 0.1.1.0