+++ ChangeLog.md
@@ -16,6 +16,7 @@
 * `--keep-going`, turned on by default for tests and benchmarks [#478](https://github.com/commercialhaskell/stack/issues/478)
 * `concurrent-tests: BOOL` [#492](https://github.com/commercialhaskell/stack/issues/492)
 * Use hashes to check file dirtiness [#502](https://github.com/commercialhaskell/stack/issues/502)
+* Install correct GHC build on systems with libgmp.so.3 [#465](https://github.com/commercialhaskell/stack/issues/465)
 
 ## 0.1.1.0